requirements.txt
flask==2.3.3
tensorflow==2.14.0
numpy==1.25.0
opencv-python==4.8.0.76
pillow==10.0.0
scikit-learn==1.3.2
matplotlib==3.8.0
